+++ CinemaApi/cinema/package.json
@@ -3,9 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/history": "^4.7.2",
+    "@types/react-router-dom": "^4.3.1",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-reveal": "^1.2.2",
+    "react-router-dom": "^4.3.1",
     "react-scripts-ts": "3.1.0"
   },
   "scripts": {
@@ -17,7 +20,7 @@
   "devDependencies": {
     "@types/jest": "^23.3.7",
     "@types/node": "^10.12.0",
-    "@types/react": "^16.4.18",
+    "@types/react": "*",
     "@types/react-dom": "^16.0.9",
     "typescript": "^3.1.3"
   }
