+++ CinemaApi/cinema/src/@Types/alltypes.d.ts
