+++ CinemaApi/cinema/src/components/Details.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import 'src/assets/css/index.css'
+class Details extends React.Component<any, any> {
+
+  public render() {
+    return (
+        <h1>Jestesmy na stronie Details</h1>
+    );
+  }
+}
+
+export default Details;
