+++ CinemaApi/cinema/src/components/NotExist.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+
+class NotExist extends React.Component<any, any> {
+
+  public render() {
+    return (
+      
+        <h1>Strona której szukasz nie istnieje!</h1>
+     
+    );
+  }
+}
+
+export default NotExist;
